neues abbruchkriterium warte maximal 1 bis 2 minuten bis das bild ausgeschnitten wurde

+++ GenderDetector/GenderDetectorControl.xaml.cs
@@ -68,6 +68,7 @@ namespace GenderDetector
 
                 // Bild speichern
                 String path = "";
+                bool imagefailed = false;
                 Dispatcher.BeginInvoke((Action)(() =>
                 {
                     CroppedBitmap colorBitmap = CropBitmap(_colorBitmap);
@@ -82,17 +83,30 @@ namespace GenderDetector
                 }));
 
                 // Warten bis Bild gespeichert wurde
-                while (path == "") { }  //TODO = nogo
-
+                int currentminute = System.DateTime.Now.Minute;
+                int diff = 0;
+                while (path == "" && !imagefailed) {
+                    diff = System.DateTime.Now.Minute - currentminute;
+                    if (diff > 2 || diff < -2) { // wenns länger als 1 min braucht brich ab
+                        imagefailed = true;
+                    }
+                }
+                Console.WriteLine(System.DateTime.Now.Minute);
                 // Auswertung des Bildes
-                CalculateGender(path);
+                if (!imagefailed)
+                {
+                    CalculateGender(path);
+                }
 
                 Dispatcher.BeginInvoke((Action)(() =>
                 {
                     // Attribute setzen
                     SetAttributes();
                     // Bild wieder löschen
-                    File.Delete(path);
+                    if (!imagefailed)
+                    {
+                        File.Delete(path);
+                    }
                 }));
             }).Start();
         }
